+++ apps/create-hypergraph/template-nextjs/Components/Login/LoginForm.tsx
@@ -15,7 +15,7 @@ export function LoginForm() {
           onClick={() => {
             redirectToConnect({
               storage: localStorage,
-              connectUrl: 'https://hypergraph-connect.vercel.app/',
+              connectUrl: 'https://connect.geobrowser.io/',
               successUrl: `${window.location.origin}/authenticate-success`,
               redirectFn: (url: URL) => {
                 window.location.href = url.toString();