+++ apps/create-hypergraph/template-vite-react/src/routes/login.tsx
@@ -14,7 +14,7 @@ function Login() {
             redirectToConnect({
               storage: localStorage,
               // connectUrl: 'http://localhost:5180',
-              connectUrl: 'https://hypergraph-connect.vercel.app/',
+              connectUrl: 'https://connect.geobrowser.io/',
               successUrl: `${window.location.origin}/authenticate-success`,
               redirectFn: (url: URL) => {
                 window.location.href = url.toString();