+++ apps/events/src/mapping.ts
@@ -6,6 +6,7 @@ export const mapping: Mapping.Mapping = {
     typeIds: [Id.Id('7f9562d4-034d-4385-bf5c-f02cdebba47a')],
     properties: {
       name: Id.Id('a126ca53-0c8e-48d5-b888-82c734c38935'),
+      description: Id.Id('9b1f76ff-9711-404c-861e-59dc3fa7d037'),
     },
     relations: {
       sponsors: Id.Id('6860bfac-f703-4289-b789-972d0aaf3abe'),