+++ apps/events/src/schema.ts
@@ -1,38 +1,38 @@
 import { Entity, Type } from '@graphprotocol/hypergraph';
 
 export class User extends Entity.Class<User>('User')({
-  name: Type.Text,
+  name: Type.String,
 }) {}
 
 export class Todo extends Entity.Class<Todo>('Todo')({
-  name: Type.Text,
-  completed: Type.Checkbox,
+  name: Type.String,
+  completed: Type.Boolean,
   assignees: Type.Relation(User),
 }) {}
 
 export class Todo2 extends Entity.Class<Todo2>('Todo2')({
-  name: Type.Text,
-  checked: Type.Checkbox,
+  name: Type.String,
+  checked: Type.Boolean,
   assignees: Type.Relation(User),
   due: Type.Date,
   amount: Type.Number,
   point: Type.Point,
-  website: Type.Text,
+  website: Type.String,
 }) {}
 
 export class JobOffer extends Entity.Class<JobOffer>('JobOffer')({
-  name: Type.Text,
+  name: Type.String,
   salary: Type.Number,
 }) {}
 
 export class Company extends Entity.Class<Company>('Company')({
-  name: Type.Text,
-  // address: Type.Text,
+  name: Type.String,
+  // address: Type.String,
   jobOffers: Type.Relation(JobOffer),
 }) {}
 
 export class Event extends Entity.Class<Event>('Event')({
-  name: Type.Text,
-  // description: Type.Text,
+  name: Type.String,
+  description: Type.optional(Type.String),
   sponsors: Type.Relation(Company),
 }) {}