+++ packages/hypergraph/src/mapping/Mapping.ts
@@ -94,7 +94,7 @@ export type SchemaDataTypeRelation = typeof SchemaDataTypeRelation.Type;
 /**
  * @since 0.2.0
  */
-export const SchemaDataTypePrimitive = EffectSchema.Literal('Text', 'Number', 'Checkbox', 'Date', 'Point');
+export const SchemaDataTypePrimitive = EffectSchema.Literal('String', 'Number', 'Boolean', 'Date', 'Point');
 /**
  * @since 0.2.0
  */
@@ -217,7 +217,7 @@ export const Schema = EffectSchema.Struct({
         {
           name: 'Account',
           knowledgeGraphId: null,
-          properties: [{ name: 'username', knowledgeGraphId: null, dataType: 'Text' }],
+          properties: [{ name: 'username', knowledgeGraphId: null, dataType: 'String' }],
         },
       ],
     },
@@ -226,7 +226,7 @@ export const Schema = EffectSchema.Struct({
         {
           name: 'Account',
           knowledgeGraphId: 'a5fd07b1-120f-46c6-b46f-387ef98396a6',
-          properties: [{ name: 'name', knowledgeGraphId: 'a126ca53-0c8e-48d5-b888-82c734c38935', dataType: 'Text' }],
+          properties: [{ name: 'name', knowledgeGraphId: 'a126ca53-0c8e-48d5-b888-82c734c38935', dataType: 'String' }],
         },
       ],
     },
@@ -771,7 +771,7 @@ export class RelationValueTypeDoesNotExistError extends Data.TaggedError(
  */
 export function mapSchemaDataTypeToGRC20PropDataType(dataType: SchemaDataType): CreatePropertyParams['dataType'] {
   switch (true) {
-    case dataType === 'Checkbox': {
+    case dataType === 'Boolean': {
       return 'CHECKBOX';
     }
     case dataType === 'Date': {