+++ packages/hypergraph/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @graphprotocol/hypergraph
 
+## 0.3.0
+### Minor Changes
+
+- cb54727: rename Type.Text to Type.String
+- c0035eb: rename Type.Checkbox to Type.Boolean
+
+### Patch Changes
+
+- f8ccaed: add Type.optional
+
 ## 0.2.0
 ### Minor Changes